x509 & ssl: add local cert retreval for ssl & x509::new_ref to support it
openssl provides an SSL_get_certificate() method to retrieve a
reference to the local certificate.

It returns this reference without incrimenting the X509's reference
count, so we need to incriment it ourselves.

To support constructing x509 objects from ffi::X509 references where we
need to incriment the ref count, provide new_ref() constructor.

new_ref() is marked unsafe due to the assumption of a non-null handle,
but other than that hazard it could be marked as safe.