(maint) Use wget with appveyor (to enable msvc testing)
This commit re-enables msvc testing by using wget with appveyor instead
of Start-FileDownload. This commit also fixes an issue where the version
of OpenSSL and Rust were specified twice in appveyor's config meaning
that it was possible to update one of the version for a release but not
the other.