Add Signer/Verifier for signing/verifying content
Previously, there was basic support for signing and verifying
signatures on buffers in PKey.  This change exposes two new
structs, `Signer` and `Verifier`, that allow for signatures
to be more easily generated/verified when performing streaming
operations.

For my initial use case, I am performing signature verification
of the entire filesystem of an embedded device which cannot fit
into RAM.

Signed-off-by: Paul Osborne <paul.osborne@smartthings.com>