signature: only allocated the required space for signature gen
Previously, we allocated a conservatively sized stack buffer
and had OpenSSL write into that before copying the contents
to an appropriately sized Vec on the heap.  OpenSSL allows for
the size of teh signature that will be generated to be determined
by calling EVP_DigestSignFinal with a NULL argument for the
destination.  We now do that which reduces the stack memory required
as well as removing the extra copy into the Vec.

Signed-off-by: Paul Osborne <paul.osborne@smartthings.com>