(maint) Recreate ability to pass in OPENSSL_LIBS variable
Prior to this commit in 43c951f743e68fac5f45119eda7c994882a1d489 the
ability to pass OPENSSL_LIBS was removed from the build.rs of
openssl-sys. This commit adds the ability to pass custom names for the
OPENSSL_LIBS back in. This is useful for when building openssl across
linux and windows with the same lib names (ssl:crypto) and the default
names provided by the build script are not valid.