Remove size check panic on Crypter::{update,finalize}()
The size of the output is deterministic for the size of the input. This means
that the caller can know when a buffer that is less than the existing check is
sufficient. This is particularly common in key wrapping scenarios.

For example, if I want to wrap two 16-byte keys using AES-CBC, the resulting
ciphertext will always be 32 bytes. Thus, I create a 32-byte buffer and call
Crypter::update() twice; once for each key. The first call succeeds, but the
second call panics. This is true even though the buffer is exactly the size
needed. Likewise, if both Crypter::update() calls succeed, the
Crypter::finalize() function will also panic.

We should, therefore, let EVP_CipherUpdate() and EVP_CipherFinal() generate any
errors when the output buffer has insufficient size and not try and be smarter
than the functions we are wrapping.