Fix static build on windows-gnu targets
Static builds for *-pc-windows-gnu targets broke, because the linker
would look for the incorrect libraries. OpenSSL builds produce
libssl.dll rather than ssl.dll which makes the linker unhappy with the
normal -lssl -lcrypto [1].

A workaround could be used:

    export OPENSSL_LIBS="ssl:crypto"

but it's arguably better to have the openssl-sys crate do the right
thing.

[1] http://www.mingw.org/wiki/specify_the_libraries_for_the_linker_to_use