Find brew openssl in correct place on Apple Silicon/aarch64
Brew recommends to keep side-by-side installations for the time being,
placing aarch64 builds into /opt/homebrew. (My feeling says using an alternative root for homebrew is there to stay,
but that may be wrong.)

That said, now `brew` in PATH may either be an aarch64 version from
`/opt/homebrew`, or it may be an intel one. And the only way I know to
differentiate them is by their path of origin and thus by convention.

Since 'pkg-config' doesn't work for me as it doesn't actually know
openssl despite it being installed as dependency and since `brew`
may refer to either architecture, hard-coding well-known paths
on MacOS seems to be the only option.