Add new feature maybe-vendored to openssl & openssl-sys
so that crates depending on openssl or openssl-sys can build
successfully even if their system does not have openssl installed.

This feature is enabled it by default since most bin crates use it without
turning on this feature and it is impossible to build these crates in
environments that do not have openssl installed, or environments that
are hard to install openssl, such as cross compilation using `cross-rs`
or `cargo-zigbuild`.

The motivation for me to submit this PR is due to
https://github.com/cargo-bins/cargo-quickinstall/issues/171

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>