Return Option from SslCipherRef::find(), check input
Previously input length was unchecked.  OpenSSL still returned NULL on a
too-short inuput, but returned a value on a too-long input.

OpenSSL doesn't add an error internally if a SSL_CIPHER_find() fails so
an Option<> seems better.

SslCipherRef::find() now checks to ensure the length of the cipher ID is
exactly 2 to avoid too-long cipher IDs matching.