Build OpenSSL et al on 32-bit x86 with -msse2
Newer BoringSSL requires SSE2 on 32-bit x86, but I opted to apply it to
all the OpenSSLs that CI builds. This allows the compiler to vectorize
random bits of the library, so hopefully it'll run faster on CI.
Especially with 32-bit x86 being so register-poor.

Additionally, Rust's definition of i686 already assumes SSE2 (see
https://github.com/rust-lang/rust/issues/82435), so there's no sense in
pretending to test pre-SSE2 targets.