openssl-sys: Provide information about the default certificate paths
The goal of these functions is to contain the logic implemented by the
`openssl-probe` crate. The `openssl-sys` crate knows whether we are
linked against the system OpenSSL library, and if so, it can use it to
get the correct certificate paths.

If we are using the vendored version, we keep the old insecure behavior
where we randomly guess paths and hope one works.