build(deps): Update `openssl-macro` to version `0.1.1`
The minimal version of `openssl-macros` needs to be the latest version. If `0.1.0` is used, then all kinds of compiler errors in the macro usage of `openssl` occur.

I found this by running `cargo minimal-versions check` on crate `native-tls`. I don't know of any simple CI check to add to this project to detect this problem in the future.