Use the new OpenSSL 3.* API for managing EVP_PKEY
The OpenSSL 3.* users now do not have a way to use non-deprecated
API by using this rust bindings, which is not sustainable in the
long term as either distributions will stop building with the
deprecated API or it will be eventually removed.

This is partially based on #2051 which was abandoned.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>