ML-KEM/ML-DSA part 2: param builder
Splitting up #2405 into a few parts as suggest by @alex.

This adds the param-builder and other openssl-sys changes.

Original commit message:

Add internal module to simplify working with OSSL_PARAM structure
We discussed that this API is not well suitable for the end users
but still, it required for several operations in OpenSSL 3.* so
instead of calling to FFI for every use of this API, this
introduces simple wrappers that allow building of the params
and their usage.

Signed-off-by: Jakub Jelen <jjelen@redhat.com>
Co-authored-by: Justus Winter <justus@sequoia-pgp.org>